backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
# Миграции
alembic==1.13.1
# Redis для кэширования
redis==5.0.1
# Структурированное логирование
structlog==24.1.0
# Тестирование
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
aiosqlite==0.20.0